Entity Classes created for diff and note table
We are now in a similar state as before with no new added functionality, but it works to start the project using the validate option

In order to start the project with the validate option enabled (default) one would need to checkout the version_1_1 branch from the database repository

+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/dsr/repository/user_login/time/TimeRepository.java
@@ -1,6 +1,6 @@
 package org.pwss.file_integrity_scanner.dsr.repository.user_login.time;
 
-import org.pwss.file_integrity_scanner.dsr.domain.user_login.entities.time.Time;
+import org.pwss.file_integrity_scanner.dsr.domain.mixed.time.Time;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -19,5 +19,5 @@ import org.springframework.stereotype.Repository;
  * @see Time
  */
 @Repository
-public interface TimeRepository extends JpaRepository<Time, Integer> {
+public interface TimeRepository extends JpaRepository<Time, Long> {
 }
