+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/dsr/service/user_login/time/TimeServiceImpl.java
@@ -1,6 +1,6 @@
 package org.pwss.file_integrity_scanner.dsr.service.user_login.time;
 
-import org.pwss.file_integrity_scanner.dsr.domain.user_login.entities.time.Time;
+import org.pwss.file_integrity_scanner.dsr.domain.mixed.time.Time;
 import org.pwss.file_integrity_scanner.dsr.repository.user_login.time.TimeRepository;
 import org.pwss.file_integrity_scanner.dsr.service.PWSSbaseService;
 import org.springframework.stereotype.Service;
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Service;
  * service implementations.
  */
 @Service
-public class TimeServiceImpl extends PWSSbaseService<TimeRepository,Time, Integer> implements TimeService {
+public class TimeServiceImpl extends PWSSbaseService<TimeRepository,Time, Long> implements TimeService {
 
    /**
      * Constructs a new instance of TimeServiceImpl with the specified repository.
